Move MG functions into MG.py
